+++ src/components/index.js
@@ -1,3 +1,3 @@
-import Icon from './Icon';
+import Icon from './Icon/Icon';
 
 export { Icon };
