+++ src/views/index.js
@@ -1,3 +1,3 @@
-import Homepage from './Homepage';
+import Homepage from './Homepage/Homepage';
 
 export { Homepage };
